requirements.txt
tomtoolkit==1.7.3
django-bootstrap4==2.2.0
django_plotly_dash==1.4.2
dash-bootstrap-components==0.10.3
whitenoise==5.1.0
dpd-static-support
django-bootstrap4
requests